# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTCAS API",
    "version": "1.0.0",
    "description": "RTCAS Backend API Definition"
  },
  "servers": [
    { "url": "http://localhost:3000" },
    { "url": "https://api.rtcas.in.th" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/admin/soldier-intake/status": {
      "patch": {
        "summary": "Update public soldier intake form status (admin)",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Status updated" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/soldier-intakes": {
      "post": {
        "summary": "Submit soldier intake (public)",
        "tags": ["Soldier Intake"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "description": "ID card image (optional)",
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/api/admin/soldier-intakes": {
      "get": {
        "summary": "List soldier intakes (company/admin)",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Delete all soldier intakes (owner)",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/soldier-intakes/export": {
      "get": {
        "summary": "Export soldier intakes as Excel",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Excel file" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/soldier-intakes/export/pdf": {
      "get": {
        "summary": "Export soldier intakes as PDF",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "PDF file" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/soldier-intakes/export/pdf/person": {
      "get": {
        "summary": "Export a soldier intake profile as PDF (by citizenId or unitCode)",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "citizenId",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "unitCode",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "PDF file" },
          "400": { "description": "Missing/invalid query" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/admin/soldier-intakes/{id}": {
      "get": {
        "summary": "Get soldier intake by id",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Record returned" },
          "400": { "description": "Invalid id" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update soldier intake by id",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "description": "ID card image (optional)",
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete soldier intake by id (owner)",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "400": { "description": "Invalid id" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/admin/soldier-intakes-summary": {
      "get": {
        "summary": "Soldier intake summary",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Summary returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/soldier-intakes/import": {
      "post": {
        "summary": "Import/update unit assignments from Excel (admin)",
        "tags": ["Soldier Intake"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Imported" },
          "400": { "description": "Invalid file or data" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  },
...
}